# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  disaster_alert_backend_internal_dto.SaveFCMTokenRequest:
    properties:
      deviceType:
        type: string
      token:
        type: string
    required:
    - token
    type: object
...
  disaster_alert_backend_internal_dto.SendTestNotificationRequest:
    properties:
      body:
        type: string
      title:
        type: string
    required:
    - body
    - title
    type: object
...
info:
  contact: {}
  description: |-
    Disaster Alert API is a mission-control backend for disaster intelligence, community incident reports, SOS escalation, assistance coordination, emergency messaging, public alerts, chats, notifications, and administrator response operations.

    ================================
    QUICK START
    ================================

    Use this Swagger page to test the API safely.

    For mobile/user endpoints:
    1. Register or login using /auth/register or /auth/login.
    2. Copy the token from the login response.
    3. Click Authorize.
    4. Paste it like this: Bearer YOUR_JWT_TOKEN.
    5. Test endpoints such as /reports, /alerts, /sos, /assistance, /notifications, and /chats.

    For admin endpoints:
    1. Enter the Admin API Key under AdminApiKeyAuth.
    2. Create or validate a privilege code.
    3. Enter the full privilege code under PrivilegeCodeAuth.
    4. Test admin endpoints such as /admin/reports, /admin/alerts, /admin/sos, /admin/assistance, and /admin/chats.

    ================================
    AUTHENTICATION GUIDE
    ================================

    1. Mobile/User endpoints use BearerAuth.
    Required header:
    Authorization: Bearer <JWT_TOKEN>

    2. Basic admin management endpoints use AdminApiKeyAuth.
    Required header:
    Sigtrack-Admin-API-Key: <ADMIN_API_KEY>

    3. Privileged admin operation endpoints require BOTH AdminApiKeyAuth and PrivilegeCodeAuth.
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
paths:
...
  /admin/notifications/test/all:
    post:
      consumes:
      - application/json
      description: |-
        Queues a broadcast test notification to all registered users.

        WHEN TO USE THIS ENDPOINT:
        Use this from the admin dashboard to verify Firebase Cloud Messaging, queue processing, and app push delivery.

        IMPORTANT:
        This endpoint should be admin-only because it sends a notification to every registered user.
        Do not expose this to normal mobile users in production.

        REQUIRED HEADERS:
        - Sigtrack-Admin-API-Key: Your admin API key.
        - X-Privilege-Code: A valid privilege code with `notifications:send`.

        REQUIRED PERMISSION:
        notifications:send

        REQUIRED FIELDS:
        - title: Notification title.
        - body: Notification message.

        EXAMPLE REQUEST BODY:
        {
        "title": "Disaster Alert Test",
        "body": "This is a test broadcast notification."
        }
      parameters:
      - description: Broadcast test notification payload. title and body are required.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disaster_alert_backend_internal_dto.SendTestNotificationRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Broadcast notification queued successfully.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or validation error.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing or invalid Admin API Key.
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Missing, revoked, expired, or unauthorized privilege code.
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Notification queue is full.
          schema:
            additionalProperties: true
            type: object
      security:
      - AdminApiKeyAuth: []
      - PrivilegeCodeAuth: []
      summary: Send notification to all users
      tags:
      - Admin Notifications
...
  /notifications/test/me:
    post:
      consumes:
      - application/json
      description: |-
        Queues a test push notification for the authenticated user.

        WHEN TO USE THIS ENDPOINT:
        Use this to confirm that the user's saved FCM token, Firebase setup, and notification queue are working.

        AUTH REQUIRED:
        This endpoint requires a valid user JWT token.
        Click Authorize and paste: Bearer YOUR_JWT_TOKEN.

        REQUIRED FIELDS:
        - title: Notification title.
        - body: Notification body.

        EXAMPLE REQUEST BODY:
        {
        "title": "Disaster Alert Test",
        "body": "This is a test notification from Disaster Alert."
        }
      parameters:
      - description: Test notification payload. title and body are required.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disaster_alert_backend_internal_dto.SendTestNotificationRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Test notification queued successfully.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or validation failed.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Notification queue is full.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Send test notification to current user
      tags:
      - User Notifications
  /notifications/token:
    post:
      consumes:
      - application/json
      description: |-
        Saves the Firebase Cloud Messaging token for the authenticated user's device.

        WHEN TO USE THIS ENDPOINT:
        The mobile app should call this after login, app start, or whenever Firebase refreshes the push token.

        WHY THIS IS IMPORTANT:
        Without saving this token, the backend may not be able to send push notifications to the user's phone.

        AUTH REQUIRED:
        This endpoint requires a valid user JWT token.
        Click Authorize and paste: Bearer YOUR_JWT_TOKEN.

        REQUIRED FIELDS:
        - token: Firebase Cloud Messaging token from the mobile app.

        OPTIONAL FIELDS:
        - deviceType: android, ios, or web.

        EXAMPLE REQUEST BODY:
        {
        "token": "fcm_token_example_abc123",
        "deviceType": "android"
        }
      parameters:
      - description: FCM token payload. token is required.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disaster_alert_backend_internal_dto.SaveFCMTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: FCM token saved successfully.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body, validation failed, or token could not
            be saved.
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Missing, invalid, or expired user token.
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error while saving FCM token.
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Save user FCM token
      tags:
      - User Notifications
...
schemes:
...
securityDefinitions:
  AdminApiKeyAuth:
    in: header
    name: Sigtrack-Admin-API-Key
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  PrivilegeCodeAuth:
    in: header
    name: X-Privilege-Code
    type: apiKey
swagger: "2.0"